cleanup tests, fix test coverage
